+++ src/vs/workbench/contrib/replNotebook/browser/replEditor.ts
@@ -60,8 +60,9 @@ import { EXECUTE_REPL_COMMAND_ID, REPL_EDITOR_ID } from 'vs/workbench/contrib/no
 import 'vs/css!./interactiveEditor';
 import { IEditorOptions } from 'vs/editor/common/config/editorOptions';
 import { deepClone } from 'vs/base/common/objects';
-import { HoverController } from 'vs/editor/contrib/hover/browser/hoverController';
 import { ReplEditorInput } from 'vs/workbench/contrib/replNotebook/browser/replEditorInput';
+import { MarginHoverController } from 'vs/editor/contrib/hover/browser/marginHoverController';
+import { ContentHoverController } from 'vs/editor/contrib/hover/browser/contentHoverController';
 
 const DECORATION_KEY = 'interactiveInputDecoration';
 const INTERACTIVE_EDITOR_VIEW_STATE_PREFERENCE_KEY = 'InteractiveEditorViewState';
@@ -377,7 +378,8 @@ export class ReplEditor extends EditorPane implements IEditorPaneWithScrolling {
 			cellEditorContributions: EditorExtensionsRegistry.getSomeEditorContributions([
 				SelectionClipboardContributionID,
 				ContextMenuController.ID,
-				HoverController.ID,
+				ContentHoverController.ID,
+				MarginHoverController.ID,
 				MarkerController.ID
 			]),
 			options: this._notebookOptions,
@@ -395,7 +397,8 @@ export class ReplEditor extends EditorPane implements IEditorPaneWithScrolling {
 					ParameterHintsController.ID,
 					SnippetController2.ID,
 					TabCompletionController.ID,
-					HoverController.ID,
+					ContentHoverController.ID,
+					MarginHoverController.ID,
 					MarkerController.ID
 				])
 			}
